Simplify the windows sdk locator interfaces
